Rename RlMixtureModel to RlSiteMixtureModel
