+++ src/revlanguage/datatypes/math/RlSiteModel.cpp
@@ -1,11 +1,3 @@
-//
-//  RlMixtureModel.cpp
-//  revbayes-proj
-//
-//  Created by Michael Landis on 3/17/15.
-//  Copyright (c) 2015 Michael Landis. All rights reserved.
-//
-
 #include <iosfwd>
 #include <string>
 #include <vector>
