+++ src/revlanguage/datatypes/math/RlSiteModel.h
@@ -2,17 +2,6 @@
  * @file
  * This file contains the declaration of the RevLanguage wrapper of the SiteModel class.
  *
- * @brief Declaration of RlSiteModel
- *
- * (c) Copyright 2009-
- * @date Last modified: $Date: 2012-08-06 20:14:22 +0200 (Mon, 06 Aug 2012) $
- * @author The RevBayes Development Core Team
- * @license GPL version 3
- * @version 1.0
- * @since 2009-11-20, version 1.0
- * @extends RbObject
- *
- * $Id: Real.h 1746 2012-08-06 18:14:22Z hoehna $
  */
 
 #ifndef RlSiteModel_H
